Use phpstan default folder structure.
