+++ src/Type/AbstractSprykerDynamicTypeExtension.php
@@ -5,7 +5,7 @@
  * Use of this software requires acceptance of the Evaluation License Agreement. See LICENSE file.
  */
 
-namespace SprykerSdk\PhpStan\DynamicType;
+namespace SprykerSdk\PhpStan\Type;
 
 use Exception;
 use PhpParser\Node\Expr\MethodCall;
