+++ src/Type/PluginDynamicTypeExtension.php
@@ -5,18 +5,21 @@
  * Use of this software requires acceptance of the Evaluation License Agreement. See LICENSE file.
  */
 
-namespace SprykerSdk\PhpStan\DynamicType;
+namespace SprykerSdk\PhpStan\Type;
 
 use PHPStan\Type\DynamicMethodReturnTypeExtension;
-use Spryker\Zed\Kernel\Persistence\AbstractQueryContainer;
+use Spryker\Zed\Kernel\Communication\AbstractPlugin;
 
-class QueryContainerDynamicTypeExtension extends AbstractSprykerDynamicTypeExtension implements DynamicMethodReturnTypeExtension
+class PluginDynamicTypeExtension extends AbstractSprykerDynamicTypeExtension implements DynamicMethodReturnTypeExtension
 {
     /**
      * @var array
      */
     protected $methodResolves = [
+        'getFacade' => true,
         'getFactory' => true,
+        'getQueryContainer' => true,
+        'getConfig' => true,
     ];
 
     /**
@@ -24,6 +27,6 @@ class QueryContainerDynamicTypeExtension extends AbstractSprykerDynamicTypeExten
      */
     public static function getClass(): string
     {
-        return AbstractQueryContainer::class;
+        return AbstractPlugin::class;
     }
 }
